pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["thiha-onow <thihaaung@onow.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
fastapi = "^0.111.0"
alembic = "^1.13.1"
sqlalchemy = "^2.0.30"
pydantic = "^2.7.3"
bcrypt = "^4.1.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-dotenv = "^1.0.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic-settings = "^2.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
